deep_desert_backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Authlib==1.2.1
python-dotenv==1.0.0 